requirements.txt
Django==2.0.3
djangorestframework==3.7.7
mysqlclient==1.3.12
pkg-resources==0.0.0
pytz==2018.3

